+++ plugins/aves_map/lib/src/interface.dart
@@ -3,7 +3,6 @@ import 'package:aves_map/src/marker/key.dart';
 import 'package:flutter/material.dart';
 import 'package:latlong2/latlong.dart';
 
-typedef ButtonPanelBuilder = Widget Function(VoidCallback resetRotation);
 typedef MarkerClusterBuilder<T> = Map<MarkerKey<T>, GeoEntry<T>> Function();
 typedef MarkerWidgetBuilder<T> = Widget Function(MarkerKey<T> key);
 typedef MarkerImageReadyChecker<T> = bool Function(MarkerKey<T> key);