+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+### Added
+
+- Map: create shortcut to custom region and filters
+
 ### Fixed
 
 - crash when loading large collection